plugins/openclaw/requirements.txt
requests>=2.28.0
watchdog>=3.0.0
websocket-client>=1.6.0
